Support release-only triplets

+++ ports/libkqueue/portfile.cmake
@@ -16,8 +16,10 @@ if(VCPKG_TARGET_IS_WINDOWS)
     file(MAKE_DIRECTORY ${CURRENT_PACKAGES_DIR}/bin)
     file(RENAME ${CURRENT_PACKAGES_DIR}/lib/kqueue.dll ${CURRENT_PACKAGES_DIR}/bin/kqueue.dll)
 
-    file(MAKE_DIRECTORY ${CURRENT_PACKAGES_DIR}/debug/bin)
-    file(RENAME ${CURRENT_PACKAGES_DIR}/debug/lib/kqueue.dll ${CURRENT_PACKAGES_DIR}/debug/bin/kqueue.dll)
+    if(NOT VCPKG_BUILD_TYPE STREQUAL "release")
+        file(MAKE_DIRECTORY ${CURRENT_PACKAGES_DIR}/debug/bin)
+        file(RENAME ${CURRENT_PACKAGES_DIR}/debug/lib/kqueue.dll ${CURRENT_PACKAGES_DIR}/debug/bin/kqueue.dll)
+    endif()
 endif()
 
 # The upstream package always creates a kqueue and kqueuestatic library; remove the static/shared library
@@ -26,9 +28,11 @@ if(VCPKG_LIBRARY_LINKAGE STREQUAL "static")
     file(REMOVE "${CURRENT_PACKAGES_DIR}/lib/kqueue.lib")
     file(RENAME "${CURRENT_PACKAGES_DIR}/lib/kqueue_static.lib" "${CURRENT_PACKAGES_DIR}/lib/kqueue.lib")
 
-    file(REMOVE "${CURRENT_PACKAGES_DIR}/debug/lib/kqueue.lib")
-    file(RENAME "${CURRENT_PACKAGES_DIR}/debug/lib/kqueue_static.lib" "${CURRENT_PACKAGES_DIR}/debug/lib/kqueue.lib")
-    
+    if(NOT VCPKG_BUILD_TYPE STREQUAL "release")
+        file(REMOVE "${CURRENT_PACKAGES_DIR}/debug/lib/kqueue.lib")
+        file(RENAME "${CURRENT_PACKAGES_DIR}/debug/lib/kqueue_static.lib" "${CURRENT_PACKAGES_DIR}/debug/lib/kqueue.lib")
+    endif()
+
     file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/bin" "${CURRENT_PACKAGES_DIR}/debug/bin")
 else()
     file(REMOVE "${CURRENT_PACKAGES_DIR}/lib/kqueue_static.lib")
